chore: move disciple modules
