+++ docs/module-structure.md
@@ -3,7 +3,7 @@
 The project organizes game logic under the `game/` directory.
 
 - **game/state.js** – central store for global state like `stats`, `systems`, `sectState`, and the current enemy. Also exposes time scaling helpers.
-- **disciple.js** – defines the `Disciple` class representing a single frog disciple.
+- **game/disciple.js** – defines the `Disciple` class representing a single frog disciple.
 - **game/disciples.js** – manages which disciples are active in combat.
 - **game/ui.js** – renders combat UI elements such as bars and disciple cards.
 - **game/combat.js** – centralizes all combat logic including damage
