+++ game/attributes.js
