+++ game/discipleAttributes.js
